Remove io.py, move readMPCORBFile to thor.utils
