+++ thor/__init__.py
@@ -3,7 +3,6 @@ from .config import *
 from .constants import *
 from .data_processing import *
 from .utils import *
-from .io import *
 from .orbits import *
 from .coordinates import *
 from .projections import *
