--- thor/tests/data/MPCORB_59580_TEST.orb
@@ -1,6 +1,0 @@
-#    Number           Ecliptic x            Ecliptic y            Ecliptic z          Ecliptic dx/dt        Ecliptic dy/dt        Ecliptic dz/dt        Epoch (TT)    Absolute   Slope
-#      or                                                                                                                                                  MJD        magnitude parameter
-#  designation           [au]                  [au]                  [au]                 [au/d]                [au/d]                [au/d]                              H         G
-#-----0001-----<>--------0039--------<>--------0040--------<>--------0041--------<>--------0042--------<>--------0043--------<>--------0044--------<>------0074-----<>---0036-<>---0037-<
-           00001  0.80405580923359E+00  0.25957854968837E+01 -0.66508145533205E-01 -0.10074377668503E-01  0.23598611144390E-02  0.19315421174722E-02   59580.17380074   3.34000  0.120000
-           00002  0.28929003638944E+01  0.21352501634356E+00 -0.38987090563201E+00 -0.34549567896537E-02  0.76533519981157E-02 -0.49985893804872E-02   59580.17380074   4.13000  0.110000
