--- thor/tests/data/MPCORB_TEST.DAT
@@ -1,45 +1,0 @@
-MINOR PLANET CENTER ORBIT DATABASE (MPCORB)
-
-This file contains published orbital elements for all numbered and unnumbered
-multi-opposition minor planets for which it is possible to make reasonable
-predictions.  It also includes published elements for recent one-opposition
-minor planets and is intended to be complete through the last issued Daily
-Orbit Update MPEC.  As such it is intended to be of interest primarily
-to astrometric observers.
-
-   Software programs may include this datafile amongst their datasets, as
-   long as this header is included (it is acceptable if it is contained
-   in a file separate from the actual data) and that proper attribution
-   to the Minor Planet Center is given.  Credit to the individual orbit
-   computers is implicit by the inclusion of a reference and the name of
-   the orbit computer on each orbit record.  Information on how to obtain
-   updated copies of the datafile must also be included.
-
-   The work of the individual astrometric observers, without whom none of
-   the work of the Minor Planet Center would be possible, is gratefully
-   acknowledged.  Credit to the individual observers is implicit by the
-   inclusion of the reference to the publication of their observations in
-   all data sets distributed by the Minor Planet Center.
-
-New versions of this file, updated on a daily basis, will be available at:
-
-          http://www.minorplanetcenter.org/iau/MPCORB/MPCORB.DAT
-
-The elements contained within MPCORB are divided into three sections,
-separated by blank lines.  The first section contains the numbered objects,
-the second section contains the unnumbered objects with perturbed orbit
-solutions and the third contains the recent 1-opposition objects with
-unperturbed orbit solutions.  Each object's elements are stored on a single
-line, the format of which is described at:
-
-          http://www.minorplanetcenter.org/iau/info/MPOrbitFormat.html
-
-If you find a problem with any data herein, please contact mpc@cfa.harvard.edu.
-
-A brief header is given below:
-
-Des'n     H     G   Epoch     M        Peri.      Node       Incl.       e            n           a        Reference #Obs #Opp    Arc    rms  Perts   Computer
-
-----------------------------------------------------------------------------------------------------------------------------------------------------------------
-00001    3.34  0.12 K1794 309.49412   73.02367   80.30888   10.59322  0.0756074  0.21408881   2.7674094  0 MPO425796  6680 114 1801-2017 0.60 M-v 30h MPCLINUX   0000      (1) Ceres              20171104
-00002    4.13  0.11 K1794 291.65136  309.99154  173.08718   34.83792  0.2305974  0.21343186   2.7730852  0 MPO421624  7910 108 1821-2017 0.58 M-v 28h MPCLINUX   0000      (2) Pallas             20171005
