--- thor/tests/data/MPCORB_TEST.orb
@@ -1,6 +1,0 @@
-#    Number           Ecliptic x            Ecliptic y            Ecliptic z          Ecliptic dx/dt        Ecliptic dy/dt        Ecliptic dz/dt        Epoch (TT)    Absolute   Slope
-#      or                                                                                                                                                  MJD        magnitude parameter
-#  designation           [au]                  [au]                  [au]                 [au/d]                [au/d]                [au/d]                              H         G
-#-----0001-----<>--------0039--------<>--------0040--------<>--------0041--------<>--------0042--------<>--------0043--------<>--------0044--------<>------0074-----<>---0036-<>---0037-<
-           00001 -0.25233197918296E+00  0.26292124489084E+01  0.12929306246756E+00 -0.10498505406583E-01 -0.17658577069768E-02  0.18798427083466E-02   58000.00000000   3.34000  0.120000
-           00002  0.23254166977602E+01  0.99298666222380E+00 -0.88089447638080E+00 -0.72293227873996E-02  0.67512559773312E-02 -0.40591166598095E-02   58000.00000000   4.13000  0.110000
