--- thor/tests/data/mpcOrbitCat_TEST.df
@@ -1,3 +1,0 @@
-designation H G epoch_pf_TT M0_deg argPeri_deg Omega_deg i_deg e n_deg_p_day a_au U ref numObs numOppos obsArc rmsResid_arcsec coarsePerturbers precisePerturbers compName flags1 flags2 readableDesignation lastObsInOrbitSolution
-00001 3.34 0.12 K1794 309.49412 73.02367 80.30888 10.59322 0.07560739999999999 0.21408881 2.7674093999999996 0 MPO425796 6680 114 1801-2017 0.6 M-v 30h MPCLINUX 0000 (1) Ceres 20171104
-00002 4.13 0.11 K1794 291.65136 309.99154 173.08718000000002 34.837920000000004 0.2305974 0.21343186 2.7730852 0 MPO421624 7910 108 1821-2017 0.58 M-v 28h MPCLINUX 0000 (2) Pallas 20171005
