--- thor/tests/data/oorbOrbitCat_TEST.df
@@ -1,3 +1,0 @@
-designation x_ec_au y_ec_au z_ec_au dx_ec/dt_au_p_day dy_ec/dt_au_p_day dz_ec/dt_au_p_day epoch_TT H G
-00001 0.8040558092335899 2.5957854968837 -0.066508145533205 -0.010074377668503 0.002359861114439 0.0019315421174722 59580.17380074 3.34 0.12
-00002 2.8929003638944 0.21352501634356 -0.38987090563201005 -0.0034549567896537 0.0076533519981157 -0.0049985893804872 59580.17380074 4.13 0.11
